Create setupHotKeysModalHotKeys.js
